backend/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
psycopg2==2.9.3